refactor: 重构冗长的agent对话界面
