+++ src/renderer/components/states/maxExpand/components/AiChatTab.tsx
@@ -27,37 +27,6 @@
 import React, { useDeferredValue, useEffect, useRef, useState, useCallback, useMemo } from 'react';
 import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
-import 'devicon/devicon.min.css';
-import javascriptOriginalIcon from 'devicon/icons/javascript/javascript-original.svg';
-import typescriptOriginalIcon from 'devicon/icons/typescript/typescript-original.svg';
-import reactOriginalIcon from 'devicon/icons/react/react-original.svg';
-import pythonOriginalIcon from 'devicon/icons/python/python-original.svg';
-import javaOriginalIcon from 'devicon/icons/java/java-original.svg';
-import cOriginalIcon from 'devicon/icons/c/c-original.svg';
-import cplusplusOriginalIcon from 'devicon/icons/cplusplus/cplusplus-original.svg';
-import csharpOriginalIcon from 'devicon/icons/csharp/csharp-original.svg';
-import goOriginalIcon from 'devicon/icons/go/go-original.svg';
-import rustOriginalIcon from 'devicon/icons/rust/rust-original.svg';
-import phpOriginalIcon from 'devicon/icons/php/php-original.svg';
-import rubyOriginalIcon from 'devicon/icons/ruby/ruby-original.svg';
-import swiftOriginalIcon from 'devicon/icons/swift/swift-original.svg';
-import kotlinOriginalIcon from 'devicon/icons/kotlin/kotlin-original.svg';
-import dartOriginalIcon from 'devicon/icons/dart/dart-original.svg';
-import htmlOriginalIcon from 'devicon/icons/html5/html5-original.svg';
-import cssOriginalIcon from 'devicon/icons/css3/css3-original.svg';
-import sassOriginalIcon from 'devicon/icons/sass/sass-original.svg';
-import lessFallbackIcon from 'devicon/icons/less/less-plain-wordmark.svg';
-import vueOriginalIcon from 'devicon/icons/vuejs/vuejs-original.svg';
-import svelteOriginalIcon from 'devicon/icons/svelte/svelte-original.svg';
-import angularOriginalIcon from 'devicon/icons/angularjs/angularjs-original.svg';
-import jsonOriginalIcon from 'devicon/icons/json/json-original.svg';
-import yamlOriginalIcon from 'devicon/icons/yaml/yaml-original.svg';
-import xmlOriginalIcon from 'devicon/icons/xml/xml-original.svg';
-import bashOriginalIcon from 'devicon/icons/bash/bash-original.svg';
-import powershellOriginalIcon from 'devicon/icons/powershell/powershell-original.svg';
-import dockerOriginalIcon from 'devicon/icons/docker/docker-original.svg';
-import sqlOriginalIcon from 'devicon/icons/azuresqldatabase/azuresqldatabase-original.svg';
-import markdownOriginalIcon from 'devicon/icons/markdown/markdown-original.svg';
 import { useTranslation } from 'react-i18next';
 import {
   resolveMihtnelisLocalToolAccess,
@@ -77,6 +46,16 @@ import { SvgIcon } from '../../../../utils/SvgIcon';
 import useIslandStore from '../../../../store/slices';
 import type { AiChatMessage, AiToolCall, AiTodoItem, AiTodoSnapshot } from '../../../../store/types';
 import { readLocalToken } from '../../../../utils/userAccount';
+import { MarkdownCodeBlock } from './agent/components/MarkdownCodeBlock';
+import { MarkdownSiteLink } from './agent/components/MarkdownSiteLink';
+import {
+  MAX_MIHTNELIS_CONTEXT_CHARS,
+  buildMihtnelisContext,
+  normalizeMarkdownCodeFences,
+  streamChatCompletion,
+  toPrettyJson,
+  unwrapJsonEnvelope,
+} from './agent/utils/chatUtils';
 
 interface ThinkEventPayload {
   text?: unknown;
@@ -139,409 +118,11 @@ const SESSION_ABORT_CONTROLLERS = new Map<string, AbortController>();
 const SESSION_STREAMING_IDS = new Set<string>();
 let cachedAiLocalToolAccessPrompt: AiLocalToolAccessPrompt | null = null;
 let cachedAiLocalToolAccessResolveError = '';
-const MAX_MIHTNELIS_CONTEXT_CHARS = 1_000_000;
 const STREAM_UI_FLUSH_INTERVAL_MS = 90;
 const VISIBLE_CHAT_WINDOW_SIZE = 4;
 const VISIBLE_CHAT_WINDOW_STEP = 4;
 const SETTINGS_OPEN_TAB_STORE_KEY = 'settings-open-tab';
 const SETTINGS_ABOUT_FEEDBACK_PREFILL_STORE_KEY = 'settings-about-feedback-prefill';
-type SiteLinkMeta = {
-  hostname: string;
-  title: string;
-  iconUrl: string;
-};
-const SITE_LINK_META_CACHE = new Map<string, SiteLinkMeta>();
-
-const CODE_LANGUAGE_ALIASES: Record<string, string> = {
-  js: 'javascript',
-  jsx: 'react',
-  ts: 'typescript',
-  tsx: 'react',
-  py: 'python',
-  yml: 'yaml',
-  sh: 'bash',
-  shell: 'bash',
-  zsh: 'bash',
-  ps1: 'powershell',
-  csharp: 'csharp',
-  cs: 'csharp',
-  'c++': 'cplusplus',
-  cpp: 'cplusplus',
-  md: 'markdown',
-};
-
-const CODE_LANGUAGE_ICON_SRC: Record<string, string> = {
-  javascript: javascriptOriginalIcon,
-  typescript: typescriptOriginalIcon,
-  react: reactOriginalIcon,
-  python: pythonOriginalIcon,
-  java: javaOriginalIcon,
-  c: cOriginalIcon,
-  cplusplus: cplusplusOriginalIcon,
-  csharp: csharpOriginalIcon,
-  go: goOriginalIcon,
-  rust: rustOriginalIcon,
-  php: phpOriginalIcon,
-  ruby: rubyOriginalIcon,
-  swift: swiftOriginalIcon,
-  kotlin: kotlinOriginalIcon,
-  dart: dartOriginalIcon,
-  html: htmlOriginalIcon,
-  css: cssOriginalIcon,
-  sass: sassOriginalIcon,
-  less: lessFallbackIcon,
-  vue: vueOriginalIcon,
-  svelte: svelteOriginalIcon,
-  angular: angularOriginalIcon,
-  json: jsonOriginalIcon,
-  yaml: yamlOriginalIcon,
-  xml: xmlOriginalIcon,
-  bash: bashOriginalIcon,
-  powershell: powershellOriginalIcon,
-  dockerfile: dockerOriginalIcon,
-  docker: dockerOriginalIcon,
-  sql: sqlOriginalIcon,
-  markdown: markdownOriginalIcon,
-};
-
-function resolveCodeLanguage(className?: string): string {
-  const match = /language-([a-zA-Z0-9#+-]+)/.exec(className ?? '');
-  const raw = (match?.[1] ?? 'plaintext').toLowerCase();
-  return CODE_LANGUAGE_ALIASES[raw] ?? raw;
-}
-
-function codeLanguageLabel(language: string): string {
-  if (language === 'plaintext') {
-    return 'TEXT';
-  }
-  return language.toUpperCase();
-}
-
-function MarkdownCodeBlock(props: { className?: string; children: React.ReactNode }): React.ReactElement {
-  const { t } = useTranslation();
-  const language = resolveCodeLanguage(props.className);
-  const iconSrc = CODE_LANGUAGE_ICON_SRC[language];
-  const code = String(props.children ?? '').replace(/\n$/, '');
-  const [copied, setCopied] = useState(false);
-
-  const handleCopy = useCallback(() => {
-    void navigator.clipboard.writeText(code).then(() => {
-      setCopied(true);
-      window.setTimeout(() => {
-        setCopied(false);
-      }, 1200);
-    }).catch(() => {});
-  }, [code]);
-
-  return (
-    <div className="max-expand-chat-code-block">
-      <div className="max-expand-chat-code-block-head">
-        <span className="max-expand-chat-code-lang">
-          {iconSrc ? (
-            <img className="max-expand-chat-code-lang-icon" src={iconSrc} alt="" aria-hidden="true" />
-          ) : (
-            <span className="max-expand-chat-code-lang-dot" aria-hidden="true" />
-          )}
-          <span>{codeLanguageLabel(language)}</span>
-        </span>
-        <button
-          type="button"
-          className="max-expand-chat-code-copy-btn"
-          onClick={handleCopy}
-        >
-          {copied
-            ? t('aiChat.codeBlock.copied', { defaultValue: '已复制' })
-            : t('aiChat.codeBlock.copy', { defaultValue: '复制' })}
-        </button>
-      </div>
-      <pre>
-        <code className={props.className}>{code}</code>
-      </pre>
-    </div>
-  );
-}
-
-function buildMihtnelisContext(messages: AiChatMessage[]): string {
-  if (!Array.isArray(messages) || messages.length === 0) {
-    return '';
-  }
-  const tailLines: string[] = [];
-  let roughLength = 0;
-  for (let i = messages.length - 1; i >= 0; i--) {
-    const msg = messages[i];
-    const role = msg?.role === 'assistant' ? 'assistant' : 'user';
-    const content = typeof msg?.content === 'string' ? msg.content.trim() : '';
-    if (!content) {
-      continue;
-    }
-    tailLines.push(`${role}: ${content}`);
-    roughLength += content.length + role.length + 4;
-    if (roughLength >= MAX_MIHTNELIS_CONTEXT_CHARS + 4096) {
-      break;
-    }
-  }
-  if (tailLines.length === 0) {
-    return '';
-  }
-  const fullContext = tailLines.reverse().join('\n\n');
-  if (fullContext.length <= MAX_MIHTNELIS_CONTEXT_CHARS) {
-    return fullContext;
-  }
-  return fullContext.slice(fullContext.length - MAX_MIHTNELIS_CONTEXT_CHARS);
-}
-
-function repairJsonControlChars(json: string): string {
-  let result = '';
-  let inString = false;
-  let escaped = false;
-  for (let i = 0; i < json.length; i++) {
-    const c = json[i];
-    if (inString) {
-      if (escaped) { result += c; escaped = false; continue; }
-      if (c === '\\') { result += c; escaped = true; continue; }
-      if (c === '"') { result += c; inString = false; continue; }
-      if (c === '\n') { result += '\\n'; continue; }
-      if (c === '\r') { result += '\\r'; continue; }
-      if (c === '\t') { result += '\\t'; continue; }
-      result += c;
-    } else {
-      result += c;
-      if (c === '"') { inString = true; }
-    }
-  }
-  return result;
-}
-
-function unwrapJsonEnvelope(content: string): string {
-  const trimmed = content.trim();
-  if (!trimmed.startsWith('{') || !trimmed.endsWith('}')) {
-    return content;
-  }
-  for (const candidate of [trimmed, repairJsonControlChars(trimmed)]) {
-    try {
-      const parsed = JSON.parse(candidate);
-      if (typeof parsed?.answer === 'string' && parsed.answer.trim()) {
-        return parsed.answer;
-      }
-    } catch { /* try next */ }
-  }
-  return content;
-}
-
-function toPrettyJson(value: unknown): string {
-  if (value == null) {
-    return '{}';
-  }
-  if (typeof value === 'string') {
-    return value;
-  }
-  try {
-    return JSON.stringify(value, null, 2);
-  } catch {
-    return String(value);
-  }
-}
-
-function normalizeMarkdownCodeFences(content: string): string {
-  if (!content || content.indexOf('```') < 0) {
-    return content;
-  }
-  let normalized = content;
-  normalized = normalized.replace(/([^\n])\s```(?=\r?\n)/g, '$1\n```');
-  normalized = normalized.replace(/([^\n])\s```\s*$/g, '$1\n```');
-  return normalized;
-}
-
-function sanitizeExternalUrl(rawUrl: string): string {
-  let value = rawUrl.trim();
-  if (!value) {
-    return '';
-  }
-  const trailingPunctuations = [',', '.', ';', '!', '?', '，', '。', '；', '！', '？'];
-  while (value.length > 0 && trailingPunctuations.includes(value[value.length - 1])) {
-    value = value.slice(0, -1);
-  }
-  while (value.endsWith(')')) {
-    const openCount = (value.match(/\(/g) ?? []).length;
-    const closeCount = (value.match(/\)/g) ?? []).length;
-    if (closeCount <= openCount) {
-      break;
-    }
-    value = value.slice(0, -1);
-  }
-  return value;
-}
-
-function buildSiteLinkMeta(url: string): SiteLinkMeta {
-  const hostname = getWebsiteHostname(url);
-  return {
-    hostname,
-    title: hostname || url,
-    iconUrl: getWebsiteFaviconUrl(url),
-  };
-}
-
-function MarkdownSiteLink(props: {
-  href: string;
-  children: React.ReactNode;
-  onClick?: React.MouseEventHandler<HTMLAnchorElement>;
-  target?: string;
-  rel?: string;
-  anchorProps?: React.AnchorHTMLAttributes<HTMLAnchorElement>;
-}): React.ReactElement {
-  const safeHref = typeof props.href === 'string' ? props.href.trim() : '';
-  const openHref = sanitizeExternalUrl(safeHref);
-  const isHttpLink = openHref.startsWith('http://') || openHref.startsWith('https://');
-  const [meta, setMeta] = useState<SiteLinkMeta>(() => {
-    if (!isHttpLink) {
-      return { hostname: '', title: '', iconUrl: '' };
-    }
-    const cached = SITE_LINK_META_CACHE.get(openHref);
-    return cached ?? buildSiteLinkMeta(openHref);
-  });
-
-  useEffect(() => {
-    if (!isHttpLink) {
-      return;
-    }
-    const cached = SITE_LINK_META_CACHE.get(openHref);
-    if (cached) {
-      setMeta(cached);
-      return;
-    }
-    const fallback = buildSiteLinkMeta(openHref);
-    setMeta(fallback);
-    let cancelled = false;
-    void fetchWebsiteTitle(openHref, 4500).then((title) => {
-      if (cancelled) {
-        return;
-      }
-      const trimmedTitle = title.trim();
-      const next: SiteLinkMeta = {
-        ...fallback,
-        title: trimmedTitle || fallback.title,
-      };
-      SITE_LINK_META_CACHE.set(openHref, next);
-      setMeta(next);
-    }).catch(() => {
-      if (!cancelled) {
-        SITE_LINK_META_CACHE.set(openHref, fallback);
-      }
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [isHttpLink, openHref]);
-
-  if (!isHttpLink) {
-    return (
-      <a
-        {...(props.anchorProps ?? {})}
-        href={props.href}
-        target={props.target ?? '_blank'}
-        rel={props.rel ?? 'noopener noreferrer'}
-        onClick={props.onClick}
-      >
-        {props.children}
-      </a>
-    );
-  }
-
-  const displayTitle = meta.title || meta.hostname || openHref;
-  const displayHost = meta.hostname;
-  const fallbackIconText = (displayTitle || '?').slice(0, 1).toUpperCase();
-
-  return (
-    <a
-      {...(props.anchorProps ?? {})}
-      href={props.href}
-      target={props.target ?? '_blank'}
-      rel={props.rel ?? 'noopener noreferrer'}
-      className="max-expand-chat-site-link"
-      title={openHref}
-      onClick={(event) => {
-        props.onClick?.(event);
-        if (event.defaultPrevented) {
-          return;
-        }
-        event.preventDefault();
-        void window.api.clipboardOpenUrl(openHref).catch(() => {});
-      }}
-    >
-      {meta.iconUrl ? (
-        <img className="max-expand-chat-site-link-icon" src={meta.iconUrl} alt="" loading="lazy" />
-      ) : (
-        <span className="max-expand-chat-site-link-icon-fallback">{fallbackIconText}</span>
-      )}
-      <span className="max-expand-chat-site-link-meta">
-        <span className="max-expand-chat-site-link-title">{displayTitle}</span>
-        {displayHost ? <span className="max-expand-chat-site-link-host">{displayHost}</span> : null}
-      </span>
-    </a>
-  );
-}
-
-/**
- * 调用 OpenAI 兼容 Chat Completions API（流式）
- */
-async function streamChatCompletion(
-  endpoint: string,
-  apiKey: string,
-  model: string,
-  messages: { role: string; content: string }[],
-  onChunk: (text: string) => void,
-  signal: AbortSignal,
-  errorMessages: {
-    apiRequestFailed: string;
-    cannotReadResponseStream: string;
-  },
-): Promise<void> {
-  const url = `${endpoint.replace(/\/+$/, '')}/chat/completions`;
-
-  const res = await fetch(url, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${apiKey}`,
-    },
-    body: JSON.stringify({ model, messages, stream: true }),
-    signal,
-  });
-
-  if (!res.ok) {
-    const body = await res.text().catch(() => '');
-    throw new Error(errorMessages.apiRequestFailed
-      .replace('{{status}}', String(res.status))
-      .replace('{{detail}}', body || res.statusText));
-  }
-
-  const reader = res.body?.getReader();
-  if (!reader) throw new Error(errorMessages.cannotReadResponseStream);
-
-  const decoder = new TextDecoder();
-  let buffer = '';
-
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) break;
-
-    buffer += decoder.decode(value, { stream: true });
-    const lines = buffer.split('\n');
-    buffer = lines.pop() || '';
-
-    for (let i = 0; i < lines.length; i++) {
-      const trimmed = lines[i].trim();
-      if (!trimmed || !trimmed.startsWith('data: ')) continue;
-      const data = trimmed.slice(6);
-      if (data === '[DONE]') return;
-      try {
-        const parsed = JSON.parse(data);
-        const delta = parsed.choices?.[0]?.delta?.content;
-        if (delta) onChunk(delta);
-      } catch { /* skip malformed chunks */ }
-    }
-  }
-}
 
 /**
  * AI 对话 Tab
